requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
channels>=4.0.0
channels-redis>=4.1.0
psycopg2-binary>=2.9.9
Pillow>=10.1.0
openpyxl>=3.1.2
reportlab>=4.0.7
python-dotenv>=1.0.0
whitenoise>=6.6.0
drf-spectacular>=0.27.0
